refactor: 테스트 코드 패키지 위치 변경 #253
